Beta 0.02 Fixed Group miss spawn. UFO now shoots rocks if no player/wedge in play. New wave does not start until UFO is out of play.

+++ Asteroids Deluxe/UFOControl.cpp
@@ -62,8 +62,7 @@ void UFOControl::SpawnUFO()
 {
 	spawnCount++;
 	ResetTimer();
-	float spawnPercent = (float)(pow(0.915, spawnCount /
-		(player->wave + 1)) * 100);
+	float spawnPercent = (powf(0.915f, (float)spawnCount / (float)(player->wave + 1)) * 100);
 
 	if (GetRandomFloat(0, 99) < spawnPercent - player->score / 500)
 	{
